Select all on file repo does not select files for clinical download (#2776)
This change restores the original behaviour for the select all button on the file repo, which is that the allRowsSelected variable is true but we don't list any selected rows. The implementation being removed here is incorrectly setting the selected rows to include only the current page of results, despite the "select all" graphics indicating the entire set is selected.

A long term solution to downloading a larger clinical data set is still required, this is simply removing a UI change that would possibly introduce confusing behaviour.